GDB-4765 Show the preview not the mapping definition configuration

+++ src/app/main/mapper/iteration/iteration.component.ts
@@ -1,4 +1,12 @@
-import {AfterViewInit, Component, Input, OnDestroy, OnInit} from '@angular/core';
+import {
+  AfterViewInit,
+  Component,
+  EventEmitter,
+  Input,
+  OnDestroy,
+  OnInit,
+  Output,
+} from '@angular/core';
 import {ModelManagementService} from 'src/app/services/model-management.service';
 import {PropertyMappingImpl} from 'src/app/models/property-mapping-impl';
 import {SimpleIRIValueMappingImpl} from 'src/app/models/simple-iri-value-mapping-impl';
@@ -12,7 +20,7 @@ import {Source} from 'src/app/models/source';
 import {ValueMappingImpl} from 'src/app/models/value-mapping-impl';
 import {MappingDetails} from 'src/app/models/mapping-details';
 import {SubjectMappingImpl} from 'src/app/models/subject-mapping-impl';
-import {BehaviorSubject, of} from 'rxjs';
+import {BehaviorSubject, Observable, of} from 'rxjs';
 import {OnDestroyMixin, untilComponentDestroyed} from '@w11k/ngx-componentdestroyed';
 import {DialogService} from 'src/app/main/components/dialog/dialog.service';
 import {TranslateService} from '@ngx-translate/core';
@@ -25,10 +33,14 @@ import {MappingBase} from 'src/app/models/mapping-base';
 import {SourceService} from 'src/app/services/source.service';
 import {MessageService} from 'src/app/services/message.service';
 import {ChannelName} from 'src/app/services/channel-name.enum';
-import {Helper} from '../../../utils/helper';
-import {plainToClass} from 'class-transformer';
-import {MapperService} from '../../../services/rest/mapper.service';
+import {Helper} from 'src/app/utils/helper';
+import {classToClass, plainToClass} from 'class-transformer';
+import {MapperService} from 'src/app/services/rest/mapper.service';
+import {ViewMode} from 'src/app/services/view-mode.enum';
 
+export interface JSONDialogData {
+  mapping
+}
 
 @Component({
   selector: 'app-iteration',
@@ -36,9 +48,11 @@ import {MapperService} from '../../../services/rest/mapper.service';
   styleUrls: ['./iteration.component.scss'],
 })
 export class IterationComponent extends OnDestroyMixin implements OnInit, AfterViewInit, OnDestroy {
-  @Input() mapping: MappingDefinitionImpl;
+  @Input() rdfMapping: Observable<MappingDefinitionImpl>;
   @Input() sources: Array<Source>;
+  @Output() updateMapping: EventEmitter<MappingDefinitionImpl> = new EventEmitter<MappingDefinitionImpl>();
 
+  mapping: MappingDefinitionImpl;
   SUBJECT = SUBJECT_SELECTOR;
   PREDICATE = PREDICATE_SELECTOR;
   OBJECT = OBJECT_SELECTOR;
@@ -49,7 +63,8 @@ export class IterationComponent extends OnDestroyMixin implements OnInit, AfterV
   usedSources: Set<string>;
 
   private boundCheckDirty: any;
-  private isPreviewOn: boolean;
+  private isPreviewOn: boolean = true;
+  private viewMode: ViewMode = ViewMode.Preview;
 
   constructor(private modelManagementService: ModelManagementService,
               public dialog: MatDialog,
@@ -64,15 +79,12 @@ export class IterationComponent extends OnDestroyMixin implements OnInit, AfterV
     super();
   }
 
-  ngOnChanges() {
-    this.initWithPreview();
-  }
-
   ngOnInit(): void {
-    this.messageService.read(ChannelName.PreviewToggle)
+    this.messageService.read(ChannelName.ViewMode)
         .pipe(untilComponentDestroyed(this))
         .subscribe((event) => {
-          this.isPreviewOn = event.getMessage();
+          this.viewMode = event.getMessage();
+          this.isPreviewOn = this.viewMode === ViewMode.Preview || this.viewMode === ViewMode.PreviewAndConfiguration;
           this.initWithPreview();
         });
 
@@ -85,6 +97,11 @@ export class IterationComponent extends OnDestroyMixin implements OnInit, AfterV
         });
 
     this.isDirty.subscribe((isDirty) => this.messageService.publish(ChannelName.DirtyMapping, isDirty));
+
+    this.rdfMapping.subscribe((mapping) => {
+      this.mapping = mapping;
+      this.init(false);
+    });
   }
 
   ngAfterViewInit() {
@@ -135,20 +152,21 @@ export class IterationComponent extends OnDestroyMixin implements OnInit, AfterV
   }
 
   initWithPreview(isDirty?: boolean) {
-    this.modelManagementService.removePreview(this.mapping);
-    if (this.mapping.getSubjectMappings().length && this.isPreviewOn && this.isComplete(this.mapping)) {
-      this.mapperService.preview(this.mapping)
+    if (this.mapping.getSubjectMappings() && this.mapping.getSubjectMappings().length && this.isPreviewOn && this.isComplete(this.mapping)) {
+      this.mapperService.preview(classToClass(this.mapping))
           .pipe(untilComponentDestroyed(this))
           .subscribe((data) => {
             this.mapping = plainToClass(MappingDefinitionImpl, data);
             this.init(isDirty);
           });
     } else {
+      this.mapping.getSubjectMappings() && this.modelManagementService.removePreview(this.mapping);
       this.init(isDirty);
     }
   }
 
   init(isDirty?: boolean) {
+    this.updateMapping.emit(this.mapping);
     if (isDirty) {
       this.isDirty.next(isDirty);
     }
@@ -307,12 +325,8 @@ export class IterationComponent extends OnDestroyMixin implements OnInit, AfterV
 
     dialogRef.afterClosed().subscribe((result) => {
       this.initMappingDetails();
-      if (result.selected === this.PREDICATE && result.mappingData.isTypeProperty) {
-        this.modelConstructService.setRootMappingInModel(result.mappingData, this.mapping);
-      } else {
-        this.modelConstructService.setRootMappingInModel(result.mappingData, this.mapping);
-        this.initWithPreview(true);
-      }
+      this.modelConstructService.setRootMappingInModel(result.mappingData, this.mapping);
+      this.initWithPreview(true);
 
       const position = result.selected === this.SUBJECT ? 1 : result.selected === this.PREDICATE ? 2 : 3;
       this.tabService.selectCommand.emit({index: this.triples.length - 2, position});
@@ -400,10 +414,14 @@ export class IterationComponent extends OnDestroyMixin implements OnInit, AfterV
     } else if (selected === this.OBJECT) {
       this.deleteObjectPropertyMapping(mapping, false);
     } else if (selected === this.PREDICATE) {
-      const propertyMappings = mapping.getSubject().getPropertyMappings();
-      const index = propertyMappings.indexOf(mapping.getPredicate());
-      if (index > -1) {
-        propertyMappings.splice(index, 1);
+      if (mapping.isTypeProperty) {
+        mapping.getSubject().setTypeMappings([]);
+      } else {
+        const propertyMappings = mapping.getSubject().getPropertyMappings();
+        const index = propertyMappings.indexOf(mapping.getPredicate());
+        if (index > -1) {
+          propertyMappings.splice(index, 1);
+        }
       }
     } else if (selected === this.SUBJECT) {
       const subject = mapping.getSubject();
@@ -486,7 +504,7 @@ export class IterationComponent extends OnDestroyMixin implements OnInit, AfterV
     window.removeEventListener('beforeunload', this.boundCheckDirty);
   }
 
-  public onValueSet(valueSet, triple: Triple, selected: string, index: number) {
+  public onValueSet(valueSet, triple: any, selected: string, index: number) {
     const value = valueSet.value;
     const source = valueSet.source;
     const previousTriple = this.triples[index - 1];
@@ -514,6 +532,7 @@ export class IterationComponent extends OnDestroyMixin implements OnInit, AfterV
         triple.setPredicate(previousTriple.getPredicate());
       }
     }
+
     let prefixTransformation;
     if (source === SourceEnum.Constant) {
       prefixTransformation = this.modelConstructService.getPrefixTransformation(value, this.getAllNamespaces());
@@ -547,7 +566,9 @@ export class IterationComponent extends OnDestroyMixin implements OnInit, AfterV
       this.modelConstructService.setCellMapping(mapping, data, settings);
       this.modelConstructService.setMappingObjectInTriple(mapping, data, settings, triple);
       this.modelConstructService.setRootMappingInModel(triple, this.mapping);
-      this.initWithPreview(true);
+      if (!!triple.getSubject() && (!!triple.getPredicate() || triple.isTypeProperty) && !!triple.getObject()) {
+        this.initWithPreview(true);
+      }
     }
   }
 
@@ -563,5 +584,9 @@ export class IterationComponent extends OnDestroyMixin implements OnInit, AfterV
     if (valueSource && valueSource.getSource() === SourceEnum.Column) {
       this.usedSources.add(valueSource.getColumnName());
     }
+  }
+
+  public getViewMode() {
+    return this.viewMode;
   }
 }
